Create DocTool to generate documentation
Move documentation generation from Tool into DocTool class in a separate
maven module to remove dependency on AsciiDoctorJ from RPM package. This
will allow to create RPM package without AsciiDoctorJ dependency, which
can be later during ovirt-engine RPM build.
DocTool will be used only in ovirt-engine-api-model to generate Ascii
and HTML documentation, but this generation happens only in CI and
generated documentation is passed into api-model RPM build as another
source archive.
