+++ doctool/src/main/java/org/ovirt/api/metamodel/doctool/AsciiDocGenerator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.ovirt.api.metamodel.tool;
+package org.ovirt.api.metamodel.doctool;
 
 import static java.util.Comparator.comparing;
 
@@ -40,6 +40,8 @@ import org.ovirt.api.metamodel.concepts.Service;
 import org.ovirt.api.metamodel.concepts.StructMember;
 import org.ovirt.api.metamodel.concepts.StructType;
 import org.ovirt.api.metamodel.concepts.Type;
+import org.ovirt.api.metamodel.tool.Names;
+import org.ovirt.api.metamodel.tool.Style;
 
 /**
  * This class takes a model and generates the corresponding AsciiDoc documentation.
