+++ doctool/src/main/java/org/ovirt/api/metamodel/doctool/AsciiDocHtmlGenerator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.ovirt.api.metamodel.tool;
+package org.ovirt.api.metamodel.doctool;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
