+++ doctool/src/main/java/org/ovirt/api/metamodel/doctool/HtmlGenerator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.ovirt.api.metamodel.tool;
+package org.ovirt.api.metamodel.doctool;
 
 /**
  * This interface is to be implemented by the classes that can translate the documentation included in the Javadoc
