+++ doctool/src/main/java/org/ovirt/api/metamodel/doctool/XmlDescriptionGenerator.java
@@ -3,21 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.ovirt.api.metamodel.tool;
+package org.ovirt.api.metamodel.doctool;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import javax.json.Json;
-import javax.json.stream.JsonGenerator;
-import javax.json.stream.JsonGeneratorFactory;
 
 import org.ovirt.api.metamodel.concepts.Annotation;
 import org.ovirt.api.metamodel.concepts.AnnotationParameter;
@@ -38,14 +29,15 @@ import org.ovirt.api.metamodel.concepts.PrimitiveType;
 import org.ovirt.api.metamodel.concepts.Service;
 import org.ovirt.api.metamodel.concepts.StructType;
 import org.ovirt.api.metamodel.concepts.Type;
+import org.ovirt.api.metamodel.runtime.xml.XmlWriter;
 
 /**
- * This class takes a model and generates its JSON description.
+ * This class takes a model and generates its XML description.
 */
 @ApplicationScoped
-public class JsonDescriptionGenerator {
+public class XmlDescriptionGenerator {
     private Model model;
-    private JsonGenerator writer;
+    private XmlWriter writer;
 
     // Reference to the object that converts documentation to HTML:
     @Inject
@@ -55,36 +47,29 @@ public class JsonDescriptionGenerator {
         // Save the model:
         this.model = model;
 
-        // Create the JSON writer:
-        try (OutputStream out = new FileOutputStream(file)) {
-            Map<String, Object> configuration = new HashMap<>();
-            configuration.put(JsonGenerator.PRETTY_PRINTING, true);
-            JsonGeneratorFactory factory = Json.createGeneratorFactory(configuration);
-            writer = factory.createGenerator(out, StandardCharsets.UTF_8);
+        // Create the XML writer:
+        try (XmlWriter tmp = new XmlWriter(file, true)) {
+            writer = tmp;
             writeModel();
-            writer.close();
-        }
-        catch (IOException exception) {
-            throw new IllegalStateException("Can't generate JSON representation", exception);
         }
     }
 
     private void writeModel() {
-        writer.writeStartObject();
+        writer.writeStartElement("model");
         Service root = model.getRoot();
         if (root != null) {
-            writer.write("root", getServiceRef(root));
+            writer.writeElement("root", getServiceRef(root));
         }
-        writer.writeStartArray("types");
+        writer.writeStartElement("types");
         model.types().forEach(this::writeType);
-        writer.writeEnd();
-        writer.writeStartArray("services");
+        writer.writeEndElement();
+        writer.writeStartElement("services");
         model.services().forEach(this::writeService);
-        writer.writeEnd();
-        writer.writeStartArray("documents");
+        writer.writeEndElement();
+        writer.writeStartElement("documents");
         model.documents().forEach(this::writeDocument);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
+        writer.writeEndElement();
     }
 
     private void writeType(Type type) {
@@ -100,107 +85,90 @@ public class JsonDescriptionGenerator {
     }
 
     private void writePrimitiveType(PrimitiveType type) {
-        writer.writeStartObject();
-        writer.write("kind", "primitive");
+        writer.writeStartElement("primitive");
         writeCommon(type);
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeStructType(StructType type) {
-        writer.writeStartObject();
-        writer.write("kind", "struct");
+        writer.writeStartElement("struct");
         writeCommon(type);
-        writer.writeStartArray("attributes");
-        type.attributes().forEach(this::writeStructAttribute);
-        writer.writeEnd();
-        writer.writeStartArray("links");
+        type.attributes().forEach(this::writeAttribute);
         type.links().forEach(this::writeStructLink);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
-    private void writeStructAttribute(Attribute attribute) {
-        writer.writeStartObject();
+    private void writeAttribute(Attribute attribute) {
+        writer.writeStartElement("attribute");
         writeCommon(attribute);
         writeTypeRef(attribute.getType());
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeStructLink(Link link) {
-        writer.writeStartObject();
+        writer.writeStartElement("link");
         writeCommon(link);
         writeTypeRef(link.getType());
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeEnumType(EnumType type) {
-        writer.writeStartObject();
-        writer.write("kind", "enum");
+        writer.writeStartElement("enum");
         writeCommon(type);
-        writer.writeStartArray("values");
         type.values().forEach(this::writeEnumValue);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeEnumValue(EnumValue value) {
-        writer.writeStartObject();
+        writer.writeStartElement("value");
         writeCommon(value);
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeService(Service service) {
-        writer.writeStartObject();
+        writer.writeStartElement("service");
         writeCommon(service);
-        writer.writeStartArray("methods");
         service.methods().forEach(this::writeServiceMethod);
-        writer.writeEnd();
-        writer.writeStartArray("locators");
         service.locators().forEach(this::writeServiceLocator);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeServiceMethod(Method method) {
-        writer.writeStartObject();
+        writer.writeStartElement("method");
         writeCommon(method);
-        writer.writeStartArray("parameters");
         method.parameters().forEach(this::writeParameter);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeServiceLocator(Locator locator) {
-        writer.writeStartObject();
+        writer.writeStartElement("locator");
         writeCommon(locator);
         writeServiceRef(locator.getService());
-        writer.writeStartArray("parameters");
         locator.parameters().forEach(this::writeParameter);
-        writer.writeEnd();
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeParameter(Parameter parameter) {
-        writer.writeStartObject();
+        writer.writeStartElement("parameter");
         writeCommon(parameter);
-        writer.write("in", parameter.isIn());
-        writer.write("out", parameter.isOut());
+        writer.writeElement("in", Boolean.toString(parameter.isIn()));
+        writer.writeElement("out", Boolean.toString(parameter.isIn()));
         writeTypeRef(parameter.getType());
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeDocument(Document document) {
-        writer.writeStartObject();
+        writer.writeStartElement("document");
         writeName(document);
         String source = document.getSource();
         if (source != null) {
-            writer.write("source", source);
+            writer.writeElement("source", source);
             String html = htmlGenerator.toHtml(source);
             if (html != null) {
-                writer.write("html", html);
+                writer.writeElement("html", html);
             }
         }
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeCommon(Concept concept) {
@@ -212,17 +180,17 @@ public class JsonDescriptionGenerator {
     private void writeName(Named named) {
         Name name = named.getName();
         if (name != null) {
-            writer.write("name", name.toString());
+            writer.writeElement("name", name.toString());
         }
     }
 
     private void writeDoc(Concept concept) {
         String doc = concept.getDoc();
         if (doc != null) {
-            writer.write("doc", doc);
+            writer.writeElement("doc", doc);
             String html = htmlGenerator.toHtml(doc);
             if (html != null) {
-                writer.write("html", html);
+                writer.writeElement("html", html);
             }
         }
     }
@@ -230,38 +198,38 @@ public class JsonDescriptionGenerator {
     private void writeAnnotations(Concept concept) {
         List<Annotation> annotations = concept.getAnnotations();
         if (!annotations.isEmpty()) {
-            writer.writeStartArray("annotations");
+            writer.writeStartElement("annotations");
             annotations.forEach(this::writeAnnotation);
-            writer.writeEnd();
+            writer.writeEndElement();
         }
     }
 
     private void writeAnnotation(Annotation annotation) {
-        writer.writeStartObject();
+        writer.writeStartElement("annotation");
         writeName(annotation);
         List<AnnotationParameter> parameters = annotation.getParameters();
         if (!parameters.isEmpty()) {
-            writer.writeStartArray("parameters");
-            parameters.forEach(this::writeAnnotationParameter);
-            writer.writeEnd();
+            writer.writeStartElement("parameters");
+            annotation.parameters().forEach(this::writeAnnotationParameter);
+            writer.writeEndElement();
         }
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeAnnotationParameter(AnnotationParameter parameter) {
-        writer.writeStartObject();
+        writer.writeStartElement("parameter");
         writeName(parameter);
         List<String> values = parameter.getValues();
         if (!values.isEmpty()) {
-            writer.writeStartArray("values");
-            values.forEach(writer::write);
-            writer.writeEnd();
+            writer.writeStartElement("values");
+            values.forEach(x -> writer.writeElement("value", x));
+            writer.writeEndElement();
         }
-        writer.writeEnd();
+        writer.writeEndElement();
     }
 
     private void writeTypeRef(Type type) {
-        writer.write("type", getTypeRef(type));
+        writer.writeElement("type", getTypeRef(type));
     }
 
     private String getTypeRef(Type type) {
@@ -277,7 +245,7 @@ public class JsonDescriptionGenerator {
     }
 
     private void writeServiceRef(Service service) {
-        writer.write("service", getServiceRef(service));
+        writer.writeElement("service", getServiceRef(service));
     }
 
     private String getServiceRef(Service service) {
