+++ doctool/src/main/resources/logging.properties
@@ -1,0 +1,8 @@
+.level=ERROR
+handlers=java.util.logging.ConsoleHandler
+java.util.logging.ConsoleHandler.level=ALL
+java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
+
+# Log record format: <LEVEL> <SOURCE> - <MESSAGE><STACKTRACE>
+java.util.logging.SimpleFormatter.format=%4$s %2$s - %5$s%6$s%n
+
