+++ ovirt-engine-api-metamodel.spec.in
@@ -29,14 +29,12 @@ BuildRequires:	mvn(org.antlr:antlr4-maven-plugin)
 BuildRequires:	mvn(org.apache.maven.plugins:maven-compiler-plugin)
 BuildRequires:	mvn(org.apache.maven.plugins:maven-source-plugin)
 BuildRequires:	mvn(org.apache.maven.plugins:maven-surefire-plugin)
-BuildRequires:	mvn(org.asciidoctor:asciidoctorj)
 BuildRequires:	mvn(org.codehaus.mojo:build-helper-maven-plugin)
 BuildRequires:	mvn(org.codehaus.mojo:exec-maven-plugin)
 BuildRequires:	mvn(org.glassfish:javax.json)
 BuildRequires:	mvn(org.jboss.resteasy:jaxrs-api)
 BuildRequires:	mvn(org.jboss.weld.se:weld-se)
 BuildRequires:	mvn(org.ovirt.maven.plugins:ovirt-jboss-modules-maven-plugin)
-BuildRequires:	mvn(org.slf4j:slf4j-api)
 BuildRequires:	mvn(org.slf4j:slf4j-jdk14)
 
 Requires:	ovirt-engine-api-metamodel-server
@@ -48,12 +46,9 @@ Requires:	mvn(commons-io:commons-io)
 Requires:	mvn(javax.json:javax.json-api)
 Requires:	mvn(javax.xml.bind:jaxb-api)
 Requires:	mvn(org.antlr:antlr4-runtime)
-Requires:	mvn(org.asciidoctor:asciidoctorj)
-Requires:	mvn(org.asciidoctor:asciidoctorj-api)
 Requires:	mvn(org.glassfish:javax.json)
 Requires:	mvn(org.jboss.spec.javax.ws.rs:jboss-jaxrs-api_2.0_spec)
 Requires:	mvn(org.jboss.weld.se:weld-se)
-Requires:	mvn(org.slf4j:slf4j-api)
 Requires:	mvn(org.slf4j:slf4j-jdk14)
 
 
@@ -98,6 +93,9 @@ Requires:	mvn(org.slf4j:slf4j-api)
 
 # no need to package tests
 %mvn_package ":metamodel-tests" __noinstall
+
+# no need to package documentation generation
+%mvn_package ":metamodel-doctool" __noinstall
 
 %if %{?skip_tests}
 # We need to skip test execution on COPR due to some weld classloading issues
